fix: add injectTraceCtx to getty
